refactor(component): debug форма вынесена в отдельный компонент
